Fix ShmError on Vagrat
since vagrant puts the localmemcache file on a shared file system this
causes mmap to fail, since a mounted shared FS does not support mmapped
files. Therefor the tests now put the shared file in `/tmp/` to be safe.

Also:

Fix bacon `at_exit` handler, `at_exit` is called again for the
`SystemExit` exception, if this is reraised it causes the script to show
an error even if the tests pass, so we `exit 0` if there has not been an
exception.